Refactor GameScene to use RootRenderNode
