+++ GameOfRunes/GameOfRunes/GameEngine/Entities/EnemyEntity.swift
@@ -15,11 +15,17 @@ class EnemyEntity: Entity {
 
     init(enemyType: EnemyType, gameEngine: GameEngine) {
         super.init()
-        
+
+        // TODO: Added by you
+        guard let renderNodeSize = gameEngine.rootRenderNode?.size else {
+            print("Unable to obtain rootRenderNode size")
+            return
+        }
+
         let enemyNode = SKSpriteNode(texture: TextureContainer.getEnemyTexture(enemyType))
         
-        let sceneSize = gameEngine.gameScene?.size ?? UIScreen.main.bounds.size
-        enemyNode.size = enemyNode.size.scaleTo(width: sceneSize.width / 6)
+//        let sceneSize = gameEngine.gameScene?.size ?? UIScreen.main.bounds.size
+        enemyNode.size = enemyNode.size.scaleTo(width: renderNodeSize.width / 6)
         CollisionType.enemyUnit.setPhysicsBody(
             for: enemyNode,
             with: enemyNode.size
